chronoctl auth whoami (#54)
Adds chronoctl auth whoami which resolves a given token to the user to which it belongs. whoami supports the standard mechanisms for providing credentials, i.e. --api-token, CHRONOSPHERE_API_TOKEN, and local token storage. If multiple credentials are present, it determines which credential to resolve in the same manner as all other chronoctl commands.